aiden-recommender/pyproject.toml
[tool.poetry]
name = "aiden-recommender"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.7.1"
requests = "^2.31.0"
selenium = "^4.20.0"
loguru = "^0.7.2"
chompjs = "^1.2.3"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
redis = "^5.0.4"
qdrant-client = "^1.9.1"
mistralai = "^0.1.8"
api-offres-emploi = "^0.0.2"
unidecode = "^1.3.8"
bs4 = "^0.0.2"
aiohttp = "^3.9.5"
jmespath = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
